Add files via upload
This is a beta version of the Shelly.clz file.  I believe it fixes a bug that was keeping the RGBW module from working with the shelly color bulb.  I only had limited time to test this so I am calling it a beta.  I also didn't have time to test whether the fix I made impacted the ability of the module to properly control the Shelly RGBW2.  Simply overwrite the version of the Shelly.clz file included in the Shelly S# v10.zip to try this version.